Fixed text rendering issues
The text wasn't rendering due to shader issues. In Renderbatch::createVertexArray, the Vertex has 2 pos for position, 3 pos for colors and 2 pos for uv in that order. But the shaders inspired from the tutorial expected a vec4 with 2 for pos and 2 for uv, so this was a mismatch. So the shaders needed to be adapted to the Renderbatch attribute structure, and new attributes had to be added in spritefont.init().